chore: move bucket config to pkg utils
