+++ pkg/config/utils.go
@@ -3,6 +3,8 @@ package config
 import (
 	"path/filepath"
 	"strings"
+
+	types "github.com/supabase/cli/pkg/storage"
 )
 
 type pathBuilder struct {
@@ -66,6 +68,23 @@ func NewPathBuilder(configPath string) pathBuilder {
 		SeedDataPath:          filepath.Join(base, "seed.sql"),
 		CustomRolesPath:       filepath.Join(base, "roles.sql"),
 	}
+}
+
+func (s *storage) GetBucketConfig() map[string]types.BucketProps {
+	result := make(map[string]types.BucketProps, len(s.Buckets))
+	for name, bucket := range s.Buckets {
+		props := types.BucketProps{
+			Public:           bucket.Public,
+			AllowedMimeTypes: bucket.AllowedMimeTypes,
+		}
+		if bucket.FileSizeLimit > 0 {
+			props.FileSizeLimit = int(bucket.FileSizeLimit)
+		} else {
+			props.FileSizeLimit = int(s.FileSizeLimit)
+		}
+		result[name] = props
+	}
+	return result
 }
 
 func ptr[T any](v T) *T {
